@@ -10,14 +10,15 @@
   "license": "ISC",
   "dependencies": {
     "axios": "~0.18.0",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
+    "babel-cli": "~6.26.0",
+    "babel-core": "~6.26.3",
     "body-parser": "~1.18.3",
     "cookie-parser": "~1.4.3",
     "csv": "~3.1.0",
-    "d3v4": "^4.2.2",
+    "d3v4": "~4.2.2",
     "domhandler": "~2.4.2",
     "domutils": "~1.7.0",
+    "eslint": "~4.19.1",
     "express": "~4.16.3",
     "express-session": "~1.15.6",
     "htmlparser": "~1.7.7",
@@ -29,27 +30,29 @@
     "passport-meetup": "~1.0.0",
     "passport-strategy": "~1.0.0",
     "request": "~2.87.0",
-    "rxjs": "~6.2.0",
+    "rxjs": "~6.2.1",
     "serve-static": "~1.13.2",
-    "ts-keycode-enum": "~1.0.4",
-    "tslib": "~1.9.1",
+    "stylelint": "~9.3.0",
+    "stylelint-config-recommended": "~2.1.0",
+    "ts-keycode-enum": "~1.0.6",
+    "tslib": "~1.9.2",
     "tslint": "~5.10.0",
     "tslint-config-standard": "~7.0.0",
     "tslint-eslint-rules": "~5.3.1",
-    "typescript": "~2.8.3"
+    "typescript": "~2.9.2"
   },
   "devDependencies": {
     "@types/body-parser": "~1.17.0",
     "@types/cookie-parser": "~1.4.1",
-    "@types/express": "~4.11.1",
-    "@types/express-session": "~1.15.8",
+    "@types/express": "~4.16.0",
+    "@types/express-session": "~1.15.10",
     "@types/htmlparser2": "~3.7.31",
-    "@types/jsdom": "~11.0.4",
-    "@types/node": "~10.1.2",
+    "@types/jsdom": "~11.0.6",
+    "@types/node": "~10.3.4",
     "@types/passport": "~0.4.5",
     "@types/passport-local": "~1.0.33",
     "@types/passport-strategy": "~0.2.33",
-    "@types/request": "~2.47.0",
+    "@types/request": "~2.47.1",
     "@types/serve-static": "~1.13.2"
   }
 }
