@@ -1,37 +1,24 @@
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
+export default class CommandInput {
+    constructor() {
+        this.subscriptions = [];
     }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports"], factory);
+    use(input) {
+        const subscriber = input.output.subscribe(this.listen && this.listen.bind(this));
+        this.subscriptions.push(subscriber);
+        return subscriber;
     }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    class CommandInput {
-        constructor() {
-            this.subscriptions = [];
-        }
-        use(input) {
-            const subscriber = input.output.subscribe(this.listen && this.listen.bind(this));
-            this.subscriptions.push(subscriber);
-            return subscriber;
-        }
-        clear() {
-            while (this.subscriptions.length) {
-                const sub = this.subscriptions.pop();
-                sub && sub.unsubscribe();
-            }
-        }
-        remove(subscription) {
-            const index = this.subscriptions.indexOf(subscription);
-            if (index > -1) {
-                subscription.unsubscribe();
-                this.subscriptions.splice(index, 1);
-            }
+    clear() {
+        while (this.subscriptions.length) {
+            const sub = this.subscriptions.pop();
+            sub && sub.unsubscribe();
         }
     }
-    exports.default = CommandInput;
-});
+    remove(subscription) {
+        const index = this.subscriptions.indexOf(subscription);
+        if (index > -1) {
+            subscription.unsubscribe();
+            this.subscriptions.splice(index, 1);
+        }
+    }
+}
 //# sourceMappingURL=command-input.js.map
