@@ -1,53 +1,40 @@
 // tslint:disable:max-classes-per-file
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
+import { Subject } from "rxjs/Subject";
+import CommandInput from "./command-input";
+import CommandOutput from "./command-output";
+export default class CommandLink {
+    constructor() {
+        this.subject = new Subject();
+        this.commandInput = new CommandInputed(this.listen.bind(this));
+        this.commandOutput = new CommandOutputed(this.subject);
     }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports", "./command-input", "./command-output", "rxjs/Subject"], factory);
+    get output() {
+        return this.commandOutput.output;
     }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    const command_input_1 = require("./command-input");
-    const command_output_1 = require("./command-output");
-    const Subject_1 = require("rxjs/Subject");
-    class CommandLink {
-        constructor() {
-            this.subject = new Subject_1.Subject();
-            this.commandInput = new CommandInputed(this.listen.bind(this));
-            this.commandOutput = new CommandOutputed(this.subject);
-        }
-        get output() {
-            return this.commandOutput.output;
-        }
-        use(input) {
-            return this.commandInput.use(input);
-        }
-        remove(subscription) {
-            return this.commandInput.remove(subscription);
-        }
-        clear() {
-            return this.commandInput.clear();
-        }
-        emit(value) {
-            this.subject && this.subject.next(value);
-        }
+    use(input) {
+        return this.commandInput.use(input);
     }
-    exports.default = CommandLink;
-    // Make a concrete version of CommandInput to be instantiated in CommandLink above
-    class CommandInputed extends command_input_1.default {
-        constructor(listener) {
-            super();
-            this.listener = listener;
-        }
-        listen(value) {
-            this.listener && this.listener(value);
-        }
+    remove(subscription) {
+        return this.commandInput.remove(subscription);
     }
-    // Make a concrete version of CommandOutput to be instantiated in CommandLink above
-    class CommandOutputed extends command_output_1.default {
+    clear() {
+        return this.commandInput.clear();
     }
-});
+    emit(value) {
+        this.subject && this.subject.next(value);
+    }
+}
+// Make a concrete version of CommandInput to be instantiated in CommandLink above
+class CommandInputed extends CommandInput {
+    constructor(listener) {
+        super();
+        this.listener = listener;
+    }
+    listen(value) {
+        this.listener && this.listener(value);
+    }
+}
+// Make a concrete version of CommandOutput to be instantiated in CommandLink above
+class CommandOutputed extends CommandOutput {
+}
 //# sourceMappingURL=command-link.js.map
