@@ -1,40 +1,40 @@
-import * as d3 from 'd3v4';
+import * as d3 from "d3v4";
 
 // Dimensions of sunburst.
-var width = 750;
-var height = 600;
-var radius = Math.min(width, height) / 2;
+const width = 750;
+const height = 600;
+const radius = Math.min(width, height) / 2;
 
 // Breadcrumb dimensions: width, height, spacing, width of tip/tail.
-var b = {
-  w: 75, h: 30, s: 3, t: 10
+const b = {
+  w: 75, h: 30, s: 3, t: 10,
 };
 
 // Mapping of step names to colors.
-var colors = {
-  "home": "#5687d1",
-  "product": "#7b615c",
-  "search": "#de783b",
-  "account": "#6ab975",
-  "other": "#a173d1",
-  "end": "#bbbbbb"
+const colors = {
+  home: "#5687d1",
+  product: "#7b615c",
+  search: "#de783b",
+  account: "#6ab975",
+  other: "#a173d1",
+  end: "#bbbbbb",
 };
 
 // Total size of all segments; we set this later, after loading the data.
-var totalSize = 0;
+let totalSize = 0;
 
-var vis = d3.select("#chart").append("svg:svg")
+const vis = d3.select("#chart").append("svg:svg")
   .attr("width", width)
   .attr("height", height)
   .append("svg:g")
   .attr("id", "container")
   .attr("transform", "translate(" + width / 2 + "," + height / 2 + ")");
 
-var partition = d3.layout.partition()
+const partition = d3.layout.partition()
   .size([2 * Math.PI, radius * radius])
   .value(function(d) { return d.size; });
 
-var arc = d3.svg.arc()
+const arc = d3.svg.arc()
   .startAngle(function(d) { return d.x; })
   .endAngle(function(d) { return d.x + d.dx; })
   .innerRadius(function(d) { return Math.sqrt(d.y); })
@@ -43,8 +43,8 @@ var arc = d3.svg.arc()
 // Use d3.text and d3.csv.parseRows so that we do not need to have a header
 // row, and can receive the csv as an array of arrays.
 d3.text("visit-sequences.csv", function(text) {
-  var csv = d3.csv.parseRows(text);
-  var json = buildHierarchy(csv);
+  const csv = d3.csv.parseRows(text);
+  const json = buildHierarchy(csv);
   createVisualization(json);
 });
 
@@ -63,12 +63,12 @@ function createVisualization(json) {
     .style("opacity", 0);
 
   // For efficiency, filter nodes to keep only those large enough to see.
-  var nodes = partition.nodes(json)
+  const nodes = partition.nodes(json)
     .filter(function(d) {
       return (d.dx > 0.005); // 0.005 radians = 0.29 degrees
     });
 
-  var path = vis.data([json]).selectAll("path")
+  const path = vis.data([json]).selectAll("path")
     .data(nodes)
     .enter().append("svg:path")
     .attr("display", function(d) { return d.depth ? null : "none"; })
@@ -83,13 +83,13 @@ function createVisualization(json) {
 
   // Get total size of the tree = value of root node from partition.
   totalSize = path.node().__data__.value;
-};
+}
 
 // Fade all but the current sequence, and show it in the breadcrumb trail.
 function mouseover(d) {
 
-  var percentage = (100 * d.value / totalSize).toPrecision(3);
-  var percentageString = percentage + "%";
+  const percentage = (100 * d.value / totalSize).toPrecision(3);
+  let percentageString = percentage + "%";
   if (percentage < 0.1) {
     percentageString = "< 0.1%";
   }
@@ -100,7 +100,7 @@ function mouseover(d) {
   d3.select("#explanation")
     .style("visibility", "");
 
-  var sequenceArray = getAncestors(d);
+  const sequenceArray = getAncestors(d);
   updateBreadcrumbs(sequenceArray, percentageString);
 
   // Fade all the segments.
@@ -141,8 +141,8 @@ function mouseleave(d) {
 // Given a node in a partition layout, return an array of all of its ancestor
 // nodes, highest first, but excluding the root.
 function getAncestors(node) {
-  var path = [];
-  var current = node;
+  const path = [];
+  let current = node;
   while (current.parent) {
     path.unshift(current);
     current = current.parent;
@@ -152,7 +152,7 @@ function getAncestors(node) {
 
 function initializeBreadcrumbTrail() {
   // Add the svg area.
-  var trail = d3.select("#sequence").append("svg:svg")
+  const trail = d3.select("#sequence").append("svg:svg")
     .attr("width", width)
     .attr("height", 50)
     .attr("id", "trail");
@@ -164,7 +164,7 @@ function initializeBreadcrumbTrail() {
 
 // Generate a string that describes the points of a breadcrumb polygon.
 function breadcrumbPoints(d, i) {
-  var points = [];
+  const points = [];
   points.push("0,0");
   points.push(b.w + ",0");
   points.push(b.w + b.t + "," + (b.h / 2));
@@ -180,12 +180,12 @@ function breadcrumbPoints(d, i) {
 function updateBreadcrumbs(nodeArray, percentageString) {
 
   // Data join; key function combines name and depth (= position in sequence).
-  var g = d3.select("#trail")
+  const g = d3.select("#trail")
     .selectAll("g")
     .data(nodeArray, function(d) { return d.name + d.depth; });
 
   // Add breadcrumb and label for entering nodes.
-  var entering = g.enter().append("svg:g");
+  const entering = g.enter().append("svg:g");
 
   entering.append("svg:polygon")
     .attr("points", breadcrumbPoints)
@@ -223,15 +223,15 @@ function updateBreadcrumbs(nodeArray, percentageString) {
 function drawLegend() {
 
   // Dimensions of legend item: width, height, spacing, radius of rounded rect.
-  var li = {
-    w: 75, h: 30, s: 3, r: 3
+  const li = {
+    w: 75, h: 30, s: 3, r: 3,
   };
 
-  var legend = d3.select("#legend").append("svg:svg")
+  const legend = d3.select("#legend").append("svg:svg")
     .attr("width", li.w)
     .attr("height", d3.keys(colors).length * (li.h + li.s));
 
-  var g = legend.selectAll("g")
+  const g = legend.selectAll("g")
     .data(d3.entries(colors))
     .enter().append("svg:g")
     .attr("transform", function(d, i) {
@@ -254,7 +254,7 @@ function drawLegend() {
 }
 
 function toggleLegend() {
-  var legend = d3.select("#legend");
+  const legend = d3.select("#legend");
   if (legend.style("visibility") == "hidden") {
     legend.style("visibility", "");
   } else {
@@ -267,24 +267,24 @@ function toggleLegend() {
 // root to leaf, separated by hyphens. The second column is a count of how
 // often that sequence occurred.
 function buildHierarchy(csv) {
-  var root = {"name": "root", "children": []};
-  for (var i = 0; i < csv.length; i++) {
-    var sequence = csv[i][0];
-    var size = +csv[i][1];
+  const root = {name: "root", children: []};
+  for (let i = 0; i < csv.length; i++) {
+    const sequence = csv[i][0];
+    const size = +csv[i][1];
     if (isNaN(size)) { // e.g. if this is a header row
       continue;
     }
-    var parts = sequence.split("-");
-    var currentNode = root;
-    for (var j = 0; j < parts.length; j++) {
-      var children = currentNode["children"];
-      var nodeName = parts[j];
-      var childNode;
+    const parts = sequence.split("-");
+    let currentNode = root;
+    for (let j = 0; j < parts.length; j++) {
+      const children = currentNode.children;
+      const nodeName = parts[j];
+      let childNode;
       if (j + 1 < parts.length) {
         // Not yet at the end of the sequence; move down the tree.
-        var foundChild = false;
-        for (var k = 0; k < children.length; k++) {
-          if (children[k]["name"] == nodeName) {
+        let foundChild = false;
+        for (let k = 0; k < children.length; k++) {
+          if (children[k].name == nodeName) {
             childNode = children[k];
             foundChild = true;
             break;
@@ -292,16 +292,16 @@ function buildHierarchy(csv) {
         }
         // If we don't already have a child node for this branch, create it.
         if (!foundChild) {
-          childNode = {"name": nodeName, "children": []};
+          childNode = {name: nodeName, children: []};
           children.push(childNode);
         }
         currentNode = childNode;
       } else {
         // Reached the end of the sequence; create a leaf node.
-        childNode = {"name": nodeName, "size": size};
+        childNode = {name: nodeName, size};
         children.push(childNode);
       }
     }
   }
   return root;
-};
+}
