@@ -1,33 +1,33 @@
-import { IMeetupUser } from './models/meetup.user';
+import { IMeetupUser } from "./models/meetup.user";
 
-console.log('start');
-import _fs from 'fs';
-import { promisify } from 'util';
+console.log("start");
+import _fs from "fs";
+import { promisify } from "util";
 // const fs = {
 //   writeFile: promisify(_fs.writeFile)
 // };
 
 // import axios from 'axios';
-import { IUserProfile, Strategy as MeetupStrategy } from 'passport-meetup';
+import { IUserProfile, Strategy as MeetupStrategy } from "passport-meetup";
 
 // import * as path from 'path';
-import * as express from 'express';
+import * as bodyParser from "body-parser";
 // import * as serveStatic from 'serve-static';
-import * as cookieParser from 'cookie-parser';
-import * as bodyParser from 'body-parser';
-import * as expressSession from 'express-session';
-import * as passport from 'passport';
-import { IMeetupEvent } from './models/meetup.events';
+import * as cookieParser from "cookie-parser";
+import * as express from "express";
+import * as expressSession from "express-session";
+import * as passport from "passport";
+import { IMeetupEvent } from "./models/meetup.events";
 
 const app = express();
 // app.use(serveStatic(path.resolve(__dirname, 'static'));
 app.use(cookieParser());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(expressSession({
-  secret: 'keyboard cat',
+  secret: "keyboard cat",
   resave: false,
   saveUninitialized: false,
-  cookie: { secure: 'auto' }
+  cookie: { secure: "auto" },
 }));
 app.use(passport.initialize());
 app.use(passport.session());
@@ -47,25 +47,25 @@ let _profile: IUserProfile;
 passport.use(new MeetupStrategy({
   },
   (token, tokenSecret, profile, done) => {
-    console.log('verify', token, tokenSecret, profile, done);
+    console.log("verify", token, tokenSecret, profile, done);
     _profile = profile;
     done(null, profile);
-  }
+  },
 ));
-app.get('/',
+app.get("/",
   async (req, res, next) => {
     const user = req && req.user;
     // console.log('/', req && req.user);
-    const urlname = 'ColoradoSprings4wheelers';
-    const statuses = ['cancelled', 'past', 'proposed', 'suggested', 'upcoming'];
+    const urlname = "ColoradoSprings4wheelers";
+    const statuses = ["cancelled", "past", "proposed", "suggested", "upcoming"];
     const params = {
-      'sign': true,
-      'key': '5d6735595632314794b567935615f74',
-      'photo-host': 'secure',
+      "sign": true,
+      "key": "5d6735595632314794b567935615f74",
+      "photo-host": "secure",
     };
-    const pars = Object.entries(params).map(param => `${param[0]}=${param[1]}`).join('&');
+    const pars = Object.entries(params).map((param) => `${param[0]}=${param[1]}`).join("&");
     const url = `https://api.meetup.com/members/self?${pars}`;
-    console.log('url', url);
+    console.log("url", url);
     const events = await axios.get(url);
     // const events = await axios.get<IMeetupEvent[]>(`https://api.meetup.com/${urlname}/events?status=${statuses.join(',')}&photo-host=secure&sign=true&key=5d6735595632314794b567935615f74`);
     const eventid = 247253189;
@@ -84,14 +84,14 @@ app.get('/',
     next();
   });
 
-app.get('/auth/meetup',
-  passport.authenticate('meetup'));
+app.get("/auth/meetup",
+  passport.authenticate("meetup"));
 
-app.get('/auth/meetup/callback',
-  passport.authenticate('meetup'),
+app.get("/auth/meetup/callback",
+  passport.authenticate("meetup"),
   (req, res) => {
     // console.log('/auth/meetup/callback', JSON.stringify(req && req.user, null, 2));
     // res.json(JSON.stringify(req && req.user, null, 2)).send();
-    res.redirect('/');
+    res.redirect("/");
   });
-console.log('end');
+console.log("end");
