@@ -1,6 +1,6 @@
-import ICommandInput from './i.command-input';
-import ICommandOutput from './i.command-output';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from "rxjs/Subscription";
+import ICommandInput from "./i.command-input";
+import ICommandOutput from "./i.command-output";
 
 export default abstract class CommandInput<InType> implements ICommandInput<InType> {
   protected subscriptions: Subscription[] = [];
