@@ -1,4 +1,4 @@
-import CommandInput from './abstract/command-input';
+import CommandInput from "./abstract/command-input";
 
 export default class OutputService extends CommandInput<string> {
   protected listen(value: string) {
