@@ -1,4 +1,4 @@
-import CommandLink from './abstract/command-link';
+import CommandLink from "./abstract/command-link";
 
 export default class StringService extends CommandLink<number, string> {
   protected listen(value: number): void {
